Fix broken test

+++ source/E2E.Tests/Services/Heroes/HeroSyncTests.cs
@@ -104,7 +104,7 @@ namespace E2E.Tests.Services.Heroes
             assertHelper.AssertPropertyOwnerField<Hero, SkillObject>(nameof(Hero._heroSkills));
             assertHelper.AssertPropertyOwnerField<Hero, CharacterAttribute>(nameof(Hero._characterAttributes));
 
-            TestEnvironment.AssertField<Hero, float>(nameof(Hero.Level), 10, defaultValue: hero.Level);
+            TestEnvironment.AssertField<Hero, int>(nameof(Hero.Level), 10, defaultValue: hero.Level);
         }
 
         // Calls the REAL patched game method (not a reflection-invoked intercept), so this covers the
